+++ package.json
@@ -52,17 +52,17 @@
   },
   "dependencies": {
     "@rifandani/nxact-yutiriti": "^1.2.2",
-    "@tanstack/vue-query": "^5.28.4",
-    "@tanstack/vue-query-devtools": "^5.28.4",
-    "@tanstack/vue-table": "^8.13.2",
+    "@tanstack/vue-query": "^5.28.7",
+    "@tanstack/vue-query-devtools": "^5.28.7",
+    "@tanstack/vue-table": "^8.15.0",
     "@vee-validate/zod": "^4.12.6",
     "@vueuse/core": "^10.9.0",
     "class-variance-authority": "^0.7.0",
     "ky": "^1.2.2",
     "pinia": "^2.1.7",
-    "radix-vue": "^1.5.2",
+    "radix-vue": "^1.5.3",
     "tailwind-merge": "^2.2.2",
-    "type-fest": "^4.12.0",
+    "type-fest": "^4.13.1",
     "v-calendar": "^3.1.2",
     "vaul-vue": "^0.1.0",
     "vee-validate": "^4.12.6",
@@ -74,36 +74,36 @@
     "zod-validation-error": "^3.0.3"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^2.8.3",
-    "@commitlint/cli": "^19.2.0",
+    "@antfu/eslint-config": "^2.9.0",
+    "@commitlint/cli": "^19.2.1",
     "@commitlint/config-conventional": "^19.1.0",
     "@iconify/vue": "^4.1.1",
     "@rollup/plugin-replace": "^5.0.5",
     "@tailwindcss/forms": "^0.5.7",
     "@tailwindcss/typography": "^0.5.10",
-    "@tanstack/eslint-plugin-query": "^5.27.7",
+    "@tanstack/eslint-plugin-query": "^5.28.6",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/user-event": "^14.5.2",
-    "@testing-library/vue": "^8.0.2",
+    "@testing-library/vue": "^8.0.3",
     "@tsconfig/node18": "^18.2.2",
     "@types/jsdom": "^21.1.6",
-    "@types/node": "^20.11.28",
+    "@types/node": "^20.11.30",
     "@vite-pwa/assets-generator": "^0.2.4",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitest/coverage-istanbul": "^1.4.0",
     "@vitest/ui": "^1.4.0",
     "@vue/tsconfig": "^0.5.1",
-    "autoprefixer": "^10.4.18",
+    "autoprefixer": "^10.4.19",
     "eslint": "^8.57.0",
     "eslint-plugin-format": "^0.1.0",
-    "eslint-plugin-jest-dom": "^5.1.0",
+    "eslint-plugin-jest-dom": "^5.2.0",
     "eslint-plugin-tailwindcss": "^3.15.1",
     "eslint-plugin-testing-library": "^6.2.0",
     "eslint-plugin-vuejs-accessibility": "^2.2.1",
     "https-localhost": "^4.7.1",
     "husky": "^9.0.11",
     "jsdom": "^24.0.0",
-    "msw": "^2.2.7",
+    "msw": "^2.2.10",
     "npm-run-all2": "^6.1.2",
     "pnpm": "^8.15.5",
     "prettier": "^3.2.5",
@@ -112,12 +112,12 @@
     "tailwindcss": "^3.4.1",
     "tailwindcss-animate": "^1.0.7",
     "ts-node": "^10.9.2",
-    "typescript": "~5.4.2",
-    "vite": "^5.1.6",
-    "vite-plugin-pwa": "^0.19.4",
+    "typescript": "~5.4.3",
+    "vite": "^5.2.4",
+    "vite-plugin-pwa": "^0.19.7",
     "vite-tsconfig-paths": "^4.3.2",
     "vitest": "^1.4.0",
-    "vue-tsc": "^2.0.6",
+    "vue-tsc": "^2.0.7",
     "workbox-core": "^7.0.0",
     "workbox-precaching": "^7.0.0",
     "workbox-routing": "^7.0.0",
